+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,8 @@
 #### 💥 Breaking
 
 - Renamed the `--upstream` option to `--remote`.
+- Removed the `--report` option from `moon check` and `moon run` commands. Reports are now always
+  created.
 
 #### 🚀 Updates